Fixing out file issues

+++ executor/redirect.c
@@ -12,19 +12,84 @@
 
 #include "../include/headers.h"
 
-void	ft_rediter(t_redirection *lst, void (*f)(t_redirection *))
+static	t_redirection	*parent_lookup(t_ast_node *ast)
 {
-	while (lst)
+	while (ast)
 	{
-		(*f)(lst);
-		if (lst->next == NULL)
+		if (ast->parent == NULL)
 			break ;
-		lst = lst->next;
+		ast = ast->parent;
+		if (ast->type == NODE_GROUP && ast->redirs)
+			return (ast->redirs);
+	}
+	return (NULL);
+}
+
+void	detect_out_redirection(t_ast_node *node)
+{
+	(void) parent_lookup;
+	if (has_outward_redirection(node))
+	{
+		char	*id = ft_itoa(node->nid);
+		char	*file = ft_strjoin("__OUTFILE__", id);
+		int 	tmp = open(file, O_WRONLY | O_CREAT, 0666);
+		free(id);
+		free(file);
+		if (tmp < 0)
+			return ;
+		node->fd = tmp;
+		if (!is_builtin(node))
+		{
+			if (dup2(tmp, STDOUT_FILENO) == -1)
+				perror("(5) Error redirecting");
+		}
 	}
 }
 
-int		has_outward_redirection(t_redirection *lst)
+void	detect_in_redirection(t_ast_node *node)
 {
+	if (has_inward_redirection(node->redirs))
+	{
+		t_redirection *lst = node->redirs;
+		char **arr = malloc(sizeof(char *) * 1);
+		arr[0] = NULL;
+		while (lst)
+		{
+			if (lst->type == REDIRECT_IN)
+				arr = add_arr_of_strs(arr, lst->file);
+			else if (lst->type == REDIRECT_HEREDOC)
+			{
+				here_doc(lst->file);
+				arr = add_arr_of_strs(arr, "__HEREDOC__");
+			}
+			if (lst->next == NULL)
+				break ;
+			lst = lst->next;
+		}
+		char *content = read_files_content(arr);
+		if (content)
+		{
+			int tmp = open("__INFILE__", O_RDWR | O_CREAT | O_TRUNC | O_DSYNC, 0666);
+			write(tmp, content, ft_strlen(content));
+			close(tmp);
+			tmp = open("__INFILE__", O_RDONLY, 0);
+
+			if (is_builtin(node))
+				node->fd = tmp;
+			else
+			{
+				if (dup2(tmp, STDIN_FILENO) == -1)
+					perror("(4) Error redirecting");
+			}
+		}
+	}
+}
+
+int		has_outward_redirection(t_ast_node *ast)
+{
+	t_redirection	*lst;
+
+	lst = ast->redirs;
 	while (lst)
 	{
 		if (lst->type == REDIRECT_OUT || lst->type == REDIRECT_APPEND)
@@ -49,7 +114,18 @@ int		has_inward_redirection(t_redirection *lst)
 	return (0);
 }
 
-static	char	*read_files_content(char **files)
+void	ft_rediter(t_redirection *lst, void (*f)(t_redirection *))
+{
+	while (lst)
+	{
+		(*f)(lst);
+		if (lst->next == NULL)
+			break ;
+		lst = lst->next;
+	}
+}
+
+char	*read_files_content(char **files)
 {
 	int		i;
 	char	*clear = NULL;
@@ -82,49 +158,6 @@ static	char	*read_files_content(char **files)
 	return (content);
 }
 
-void	detect_out_redirection(t_ast_node *node)
-{
-	if (has_outward_redirection(node->redirs))
-	{
-		int tmp = open("__OUTFILE__", O_WRONLY | O_CREAT, 0666);
-		if (dup2(tmp, STDOUT_FILENO) == -1)
-			perror("(5) Error redirecting");
-	}
-}
-
-void	detect_in_redirection(t_ast_node *node)
-{
-	if (has_inward_redirection(node->redirs))
-	{
-		t_redirection *lst = node->redirs;
-		char **arr = malloc(sizeof(char *) * 1);
-		arr[0] = NULL;
-		while (lst)
-		{
-			if (lst->type == REDIRECT_IN)
-				arr = add_arr_of_strs(arr, lst->file);
-			else if (lst->type == REDIRECT_HEREDOC)
-			{
-				here_doc(lst->file);
-				arr = add_arr_of_strs(arr, "__HEREDOC__");
-			}
-			if (lst->next == NULL)
-				break ;
-			lst = lst->next;
-		}
-		char *content = read_files_content(arr);
-		if (content)
-		{
-			int tmp = open("__INFILE__", O_RDWR | O_CREAT | O_TRUNC | O_DSYNC, 0666);
-			write(tmp, content, ft_strlen(content));
-			close(tmp);
-			tmp = open("__INFILE__", O_RDONLY, 0);
-			if (dup2(tmp, STDIN_FILENO) == -1)
-				perror("(4) Error redirecting");
-		}
-	}
-}
-
 void	pipex_redirect(t_ast_node *node, int fd[2], int files[3], int is_last)
 {
 	if (files[0] == -1)
@@ -144,7 +177,7 @@ void	pipex_redirect(t_ast_node *node, int fd[2], int files[3], int is_last)
 		close(fd[0]);
 	}
 
-	if (has_outward_redirection(node->redirs))
+	if (has_outward_redirection(node))
 	{
 		detect_out_redirection(node);
 		close(fd[0]);
@@ -191,27 +224,31 @@ void	multiple_output_redirections(t_ast_node *node)
 {
 	char	**arr;
 	char	*content;
-	char	*out;
 
-	arr = malloc(sizeof(char *) * 2);
-	out = ft_strdup("__OUTFILE__");
-	if (arr)
+	if (has_outward_redirection(node))
 	{
-		arr[0] = ft_strdup(out);
-		arr[1] = NULL;
-		content = read_files_content(arr);
-		clear_arr_of_strs(arr);
-		if (content)
+		arr = malloc(sizeof(char *) * 2);
+		char	*id = ft_itoa(node->nid);
+		char	*out = ft_strjoin("__OUTFILE__", id);
+		if (arr)
 		{
-			redlist_out(node->redirs, content);
-			free(content);
+			arr[0] = ft_strdup(out);
+			arr[1] = NULL;
+			content = read_files_content(arr);
+			clear_arr_of_strs(arr);
+			if (content)
+			{
+				redlist_out(node->redirs, content);
+				free(content);
+			}
 		}
+		if (access(out, F_OK) == 0)
+		{
+			unlink(out);
+		}
+		free(out);
+		free(id);
 	}
-	if (access(out, F_OK) == 0)
-	{
-		unlink(out);
-	}
-	free(out);
 }
 
 void	here_doc(char *delimit)
